fix assistant page name
