+++ src/pages/assistant.tsx
@@ -7,6 +7,7 @@ import {
   FormGroup,
   Grid,
   IconButton,
+  /* ThemeProvider */
   Modal,
   Skeleton,
   Stack,
@@ -15,29 +16,31 @@ import {
   useMediaQuery,
   useTheme,
 } from "@mui/material";
+// const Values = React.lazy(() => import("./modules/views/Values"));
+
 const Values = dynamic(() => import("../components/assistant/Why"), { suspense: true, ssr: false });
+
 const WhoWeAre = dynamic(() => import("../components/home/views/WhoWeAre"), { suspense: true, ssr: false });
 
 import dynamic from "next/dynamic";
-import { useRouter } from "next/router";
 import React, { ReactNode, Suspense, useCallback, useEffect, useRef, useState } from "react";
 import { Rive, useRive } from "rive-react";
 
-import PublicLayout from "@/components/layouts/PublicLayout";
 import SearcherPupUp from "@/components/SearcherPupUp";
+import { ThemeSwitcher } from "@/components/ThemeSwitcher";
 import { useInView } from "@/hooks/useObserver";
-// import useThemeChange from "@/hooks/useThemeChange";
+import useThemeChange from "@/hooks/useThemeChange";
 import { useWindowSize } from "@/hooks/useWindowSize";
 
+// import { brandingDarkTheme } from "@/lib/theme/brandingTheme";
 import LogoDarkMode from "../../public/DarkModeLogoMini.png";
-import AppFooter from "../components/AppFooter2";
+import AppFooter from "../components/AppFooter2"; // TODO: load with lazy load and observer when is required
 import AppHeaderSearchBar from "../components/AppHeaderSearchBar";
 import HowItWorks from "../components/assistant/HowItWorks";
 import { sectionsOrder } from "../components/assistant/sectionsOrder";
 import { MemoizedTableOfContent } from "../components/home/components/TableOfContent";
 import CustomTypography from "../components/home/components/Typography";
-import { ThemeSwitcher } from "../components/ThemeSwitcher";
-import useThemeChange from "../hooks/useThemeChange";
+import PublicLayout from "../components/layouts/PublicLayout";
 
 /**
  * animations builded with: https://rive.app/
@@ -51,9 +54,7 @@ export const gray03 = "#AAAAAA";
 const section1ArtBoards = [
   { name: "artboard-1", durationMs: 1000, getHeight: (vh: number) => vh - HEADER_HEIGTH, color: "#ff28c9" },
 ];
-
 const artboards = [
-  // { name: "Hero", durationMs: 1000, getHeight: (vh: number) => vh, color: "#4eb8f5" },
   { name: "Planning", durationMs: 17000, getHeight: (vh: number) => 6 * vh, color: "#f33636" },
   { name: "Meetings", durationMs: 24000, getHeight: (vh: number) => 8 * vh, color: "#f38b36" },
   { name: "Goals", durationMs: 40000, getHeight: (vh: number) => 15 * vh, color: "#e6f336" },
@@ -62,29 +63,23 @@ const artboards = [
 export const SECTION_WITH_ANIMATION = 1;
 
 const sectionsTmp = [
+  { id: "Landing", title: "Home", simpleTitle: "Home", children: [] },
   {
-    id: "Hero Section",
-    title: "Hero Section",
-    simpleTitle: "Hero",
-    children: [],
-    hidden: true,
-  },
-  {
-    id: "HowItWorksSection",
+    id: "How1CademyAssistantWorks",
     title: "How 1Cademy Assistant works?",
     simpleTitle: "How?",
     children: [
-      // { id: "animation1", title: "Planning", simpleTitle: "Planning" },
       { id: "animation1", title: "Planning", simpleTitle: "Planning" },
       { id: "animation2", title: "Meetings", simpleTitle: "Meetings" },
       { id: "animation3", title: "Goals", simpleTitle: "Goals" },
     ],
   },
-  { id: "ValuesSection", title: "Why 1Cademy Assistant?", simpleTitle: "Why?", children: [] },
-  { id: "WhoWeAreSection", title: "Who's Behind 1Cademy Assistant?", simpleTitle: "Who?", children: [] },
+  { id: "Why1CademyAssistant", title: "Why 1Cademy Assistant?", simpleTitle: "Why?", children: [] },
+  { id: "WhoIsBehind1CademyAssistant", title: "Who's Behind 1Cademy Assistant?", simpleTitle: "Who?", children: [] },
 ];
 
 const Home = () => {
+  // const [section, setSection] = useState(0);
   const theme = useTheme();
 
   const [sectionSelected, setSelectedSection] = useState(0);
@@ -95,12 +90,12 @@ const Home = () => {
   const isMovil = useMediaQuery("(max-width:600px)");
   const isTablet = useMediaQuery("(min-width:900px)");
   const [showLandingOptions, setShowLandingOptions] = useState(true);
-  const [animationSelected, setSelectedAnimation] = useState(-1);
+  // const [/* showAnimationOptions, */ setShowAnimationOptions] = useState(false);
+  const [animationSelected, setSelectedAnimation] = useState(0);
   const [handleThemeSwitch] = useThemeChange();
 
   const [open, setOpen] = useState(false);
   const [openSearch, setOpenSearch] = useState(false);
-  const router = useRouter();
 
   const { entry: whyEntry, inViewOnce: whyInViewOnce, ref: whySectionRef } = useInView();
   const { entry: whoEntry, inViewOnce: whoInViewOnce, ref: whoSectionRef } = useInView();
@@ -109,27 +104,29 @@ const Home = () => {
   const HomeSectionRef = useRef<HTMLDivElement | null>(null);
   const howSectionRef = useRef<HTMLDivElement | null>(null);
   const timeInSecondsRef = useRef<number>(0);
-  const { rive: rive0, RiveComponent: RiveComponent0 } = useRive({
+
+  const { rive: rive1, RiveComponent: RiveComponent1 } = useRive({
     src: "rive-assistant/assistant-0.riv",
     artboard: "New Artboard",
     animations: ["Timeline 1", "eyes"],
     autoplay: true,
   });
-  const { rive: rive1, RiveComponent: RiveComponent1 } = useRive({
+
+  const { rive: rive2, RiveComponent: RiveComponent2 } = useRive({
     src: "rive-assistant/assistant-1.riv",
     artboard: "artboard-1",
     animations: ["Timeline 1", "dark", "light"],
     autoplay: false,
   });
 
-  const { rive: rive2, RiveComponent: RiveComponent2 } = useRive({
+  const { rive: rive3, RiveComponent: RiveComponent3 } = useRive({
     src: "rive-assistant/assistant-2.riv",
     artboard: "artboard-2",
     animations: ["Timeline 1", "dark", "light"],
     autoplay: false,
   });
 
-  const { rive: rive3, RiveComponent: RiveComponent3 } = useRive({
+  const { rive: rive4, RiveComponent: RiveComponent4 } = useRive({
     src: "rive-assistant/assistant-3.riv",
     artboard: "artboard-3",
     animations: ["Timeline 1", "dark", "light"],
@@ -137,12 +134,13 @@ const Home = () => {
   });
 
   useEffect(() => {
-    if (!rive1 || !rive2 || !rive3) return;
+    if (!rive1 || !rive2 || !rive3 || !rive4) return;
 
     advanceAnimationTo(rive1, timeInSecondsRef.current, theme);
     advanceAnimationTo(rive2, timeInSecondsRef.current, theme);
     advanceAnimationTo(rive3, timeInSecondsRef.current, theme);
-  }, [rive1, rive2, rive3, theme]);
+    advanceAnimationTo(rive4, timeInSecondsRef.current, theme);
+  }, [rive1, rive2, rive3, rive4, theme]);
 
   useEffect(() => {
     const hash = window?.location?.hash;
@@ -150,22 +148,23 @@ const Home = () => {
 
     const selectedSectionByUrl = sectionsTmp.findIndex(cur => `#${cur.id}` === hash);
     if (selectedSectionByUrl < 0) return;
+
     setSelectedSection(selectedSectionByUrl);
   }, []);
 
   const getSectionHeights = useCallback(() => {
-    if (!howSectionRef?.current) return null;
     if (!HomeSectionRef?.current) return null;
+    if (!howSectionRef?.current) return null;
     if (!whyEntry) return null;
     if (!whoEntry) return null;
 
     return [
       { id: HomeSectionRef.current.id, height: 0 },
-      { id: howSectionRef.current.id, height },
-      { id: whyEntry.target.id, height: howSectionRef.current.clientHeight },
+      { id: howSectionRef.current.id, height: HomeSectionRef.current.clientHeight },
+      { id: whyEntry.target.id, height: howSectionRef.current.clientHeight - HomeSectionRef.current.clientHeight },
       { id: whoEntry.target.id, height: whyEntry.target.clientHeight },
     ];
-  }, [height, whoEntry, whyEntry]);
+  }, [whoEntry, whyEntry]);
 
   const getSectionPositions = useCallback(() => {
     if (!HomeSectionRef?.current) return null;
@@ -174,13 +173,10 @@ const Home = () => {
     if (!whoEntry) return null;
 
     return [
-      {
-        id: HomeSectionRef.current.id,
-        height: HomeSectionRef.current.clientHeight,
-      },
+      { id: HomeSectionRef.current.id, height: HomeSectionRef.current.clientHeight },
       {
         id: howSectionRef.current.id,
-        height: howSectionRef.current.clientHeight,
+        height: howSectionRef.current.clientHeight - HomeSectionRef.current.clientHeight,
       },
       { id: whyEntry.target.id, height: whyEntry.target.clientHeight },
       { id: whoEntry.target.id, height: whoEntry.target.clientHeight },
@@ -210,25 +206,22 @@ const Home = () => {
     (
       event: any,
       {
-        rive0,
         rive1,
         rive2,
         rive3,
+        rive4,
       }: {
-        rive0: Rive | null;
         rive1: Rive | null;
         rive2: Rive | null;
         rive3: Rive | null;
+        rive4: Rive | null;
       }
     ) => {
-      console.log("detect");
-      if (!rive1 || !rive2 || !rive3 || !rive0) return;
-      console.log("detect 1");
+      if (!rive1 || !rive2 || !rive3 || !rive4) return;
       if (notSectionSwitching) {
-        console.log("detect 2");
         const currentScrollPosition = event.target.scrollTop;
         const sectionsHeight = getSectionPositions();
-        console.log({ sectionsHeight });
+        // console.log({ sectionsHeight });
         if (!sectionsHeight) return;
 
         const { min, idx: idxSection } = sectionsHeight.reduce(
@@ -238,18 +231,15 @@ const Home = () => {
           },
           { max: 0, min: 0, idx: -1 }
         );
-        console.log(1);
+
         if (idxSection < 0) return;
-        console.log(2);
+
         let animationsHeight = [];
         if (idxSection === 0) {
           animationsHeight = [section1ArtBoards[0].getHeight(height)];
         } else {
           animationsHeight = getAnimationsPositions();
         }
-        // animationsHeight = getAnimationsPositions();
-
-        // console.log({ animationsHeight });
 
         const { maxAnimation, minAnimation, idxAnimation } = animationsHeight.reduce(
           (acu, cur, idx) => {
@@ -264,23 +254,23 @@ const Home = () => {
         if (window.location.hash !== `#${sectionSelected.id}`) {
           window.history.replaceState(null, sectionSelected.title, "#" + sectionSelected.id);
         }
-
         setSelectedSection(idxSection);
         setSelectedAnimation(idxAnimation);
 
         let showLandingOptions = false;
+        // let showEndAnimationOptions = false;
 
         if (idxAnimation < 0) return;
 
-        console.log({ idxSection });
         if (idxSection === 0) {
           const lowerAnimationLimit = minAnimation;
           const upperAnimationLimit = maxAnimation;
           const rangeFrames = upperAnimationLimit - lowerAnimationLimit;
           const positionFrame = currentScrollPosition - lowerAnimationLimit;
           const percentageFrame = (positionFrame * 100) / rangeFrames;
+          setIdxRiveComponent(0);
           // if (percentageFrame < 50) {
-          //   setIdxRiveComponent(0);
+          //  setIdxRiveComponent(0);
           // } else {
           //   const newLowerAnimationLimit = lowerAnimationLimit + rangeFrames / 2;
           //   const newPositionFrame = currentScrollPosition - newLowerAnimationLimit;
@@ -292,10 +282,9 @@ const Home = () => {
           //   setIdxRiveComponent(1);
           // }
 
-          if (percentageFrame < 2) {
+          if (percentageFrame < 5) {
             showLandingOptions = true;
           }
-          setIdxRiveComponent(0);
         }
 
         if (idxSection === SECTION_WITH_ANIMATION) {
@@ -310,33 +299,25 @@ const Home = () => {
           const timeInSeconds = (artboards[idxAnimation].durationMs * percentageFrame) / (1000 * 100);
           timeInSecondsRef.current = timeInSeconds;
 
-          //   console.log({ timeInSecondsRef: timeInSecondsRef.current.toFixed(2) });
-
           if (idxAnimation === 0) {
-            advanceAnimationTo(rive1, timeInSeconds, theme);
-          }
-          if (idxAnimation === 1) {
             advanceAnimationTo(rive2, timeInSeconds, theme);
           }
-          if (idxAnimation === 2) {
+          if (idxAnimation === 1) {
             advanceAnimationTo(rive3, timeInSeconds, theme);
           }
-          //   if (idxAnimation === 1) {
-          //     advanceAnimationTo(rive4, timeInSeconds, theme);
-          //   }
-          //   if (idxAnimation === 2) {
-          //     advanceAnimationTo(rive5, timeInSeconds, theme);
-          //   }
-          //   if (idxAnimation === 3) {
-          //     advanceAnimationTo(rive6, timeInSeconds, theme);
-          //     if (percentageFrame > 50) {
-          //       showEndAnimationOptions = true;
-          //     }
-          //   }
+          if (idxAnimation === 2) {
+            advanceAnimationTo(rive4, timeInSeconds, theme);
+            // if (percentageFrame > 50) {
+            //   showEndAnimationOptions = true;
+            // }
+          }
+          // if (idxAnimation === 3) {
+          //   advanceAnimationTo(rive5, timeInSeconds, theme);
+
+          // }
         }
 
         // update options display
-
         setShowLandingOptions(showLandingOptions);
       }
     },
@@ -345,22 +326,18 @@ const Home = () => {
 
   const switchSection = useCallback(
     (sectionIdx: number, animationIndex = 0) => {
-      if (!rive1 || !rive2 || !rive3 || !rive0) return;
+      if (!rive2 || !rive3 || !rive4) return;
 
-      console.log("---->", { sectionIdx, animationIndex });
       setNotSectionSwitching(false);
       const sectionsHeight = getSectionHeights();
       if (!sectionsHeight) return;
 
-      console.log({ sectionsHeight });
-
       const previousSections = sectionsHeight.slice(0, sectionIdx + 1);
       const sectionResult = previousSections.reduce((a, c) => ({ id: c.id, height: a.height + c.height }));
-      //   console.log({ sectionIdx, animationIndex });
 
       let cumulativeAnimationHeight = 0;
+
       const animationsHeight = getAnimationsHeight();
-      console.log({ sectionIdx, animationIndex, animationsHeight });
       if (animationsHeight) {
         const previousAnimationHeight = animationsHeight.slice(0, animationIndex + 1);
         cumulativeAnimationHeight = previousAnimationHeight.reduce((a, c) => a + c);
@@ -371,25 +348,20 @@ const Home = () => {
       // setSelectedSection(sectionIdx);
       if (sectionIdx === 0) {
         setShowLandingOptions(true);
+        setIdxRiveComponent(animationIndex);
       }
       if (sectionIdx === SECTION_WITH_ANIMATION) {
-        setIdxRiveComponent(animationIndex);
+        setIdxRiveComponent(animationIndex + 1);
         // reset animation when jump through sections
         if (animationIndex === 0) {
-          rive0.scrub("Timeline 1", 0);
-        }
-        if (animationIndex === 1) {
-          rive1.scrub("Timeline 1", 0);
-        }
-        if (animationIndex === 2) {
           rive2.scrub("Timeline 1", 0);
         }
-        if (animationIndex === 3) {
+        if (animationIndex === 1) {
           rive3.scrub("Timeline 1", 0);
         }
-        // if (animationIndex === 2) {
-        //   rive5.scrub("Timeline 1", 0);
-        // }
+        if (animationIndex === 2) {
+          rive4.scrub("Timeline 1", 0);
+        }
         // if (animationIndex === 3) {
         //   rive6.scrub("Timeline 1", 0);
         // }
@@ -402,18 +374,13 @@ const Home = () => {
         setNotSectionSwitching(true);
       }, 1000);
     },
-    [getAnimationsHeight, getSectionHeights, rive0, rive1, rive2, rive3]
+    [getAnimationsHeight, getSectionHeights, rive2, rive3, rive4]
   );
 
-  //   const signUpHandler = () => {
-  //     router.push("/signin");
-  //   };
-
   return (
-    // <ThemeProvider theme={brandingLightTheme}>
     <Box
       id="ScrollableContainer"
-      onScroll={e => detectScrollPosition(e, { rive1, rive2, rive3, rive0 })}
+      onScroll={e => detectScrollPosition(e, { rive1, rive2, rive3, rive4 })}
       sx={{
         height: "100vh",
         overflowY: "auto",
@@ -433,16 +400,13 @@ const Home = () => {
             width: "100%",
             background: theme => (theme.palette.mode === "dark" ? "rgba(0,0,0,.72)" : "#f8f8f894"),
             backdropFilter: "saturate(180%) blur(20px)",
-
-            // filter: "blur(1px)"
           }}
-          // style={{willChange:"filter"}}
         />
         <Stack
           direction={"row"}
           justifyContent={"space-between"}
           alignItems={"center"}
-          spacing={"20px"}
+          spacing={isMovil ? "10px" : "20px"}
           sx={{
             width: "100%",
             maxWidth: "980px",
@@ -458,9 +422,7 @@ const Home = () => {
             direction={"row"}
             sx={{ color: "#f8f8f8" }}
           >
-            <Box onClick={() => router.push("/")} sx={{ cursor: "pointer" }}>
-              <img src={LogoDarkMode.src} alt="logo" width="45px" height={"45px"} />
-            </Box>
+            <img src={LogoDarkMode.src} alt="logo" width="45px" height={"45px"} />
 
             {isTablet && (
               <>
@@ -469,11 +431,11 @@ const Home = () => {
                     sx={{
                       cursor: "pointer",
                       borderBottom: theme =>
-                        sectionSelected === 0 ? `solid 2px ${theme.palette.common.orange}` : undefined,
+                        sectionSelected === 1 ? `solid 2px ${theme.palette.common.orange}` : undefined,
                     }}
                     onClick={() => switchSection(0)}
                   >
-                    {sectionsOrder[0].label}
+                    {sectionsOrder[1].label}
                   </Typography>
                 </Tooltip>
                 <Tooltip title={sectionsOrder[1].title}>
@@ -481,11 +443,11 @@ const Home = () => {
                     sx={{
                       cursor: "pointer",
                       borderBottom: theme =>
-                        sectionSelected === 1 ? `solid 2px ${theme.palette.common.orange}` : undefined,
+                        sectionSelected === 2 ? `solid 2px ${theme.palette.common.orange}` : undefined,
                     }}
                     onClick={() => switchSection(1)}
                   >
-                    {sectionsOrder[1].label}
+                    {sectionsOrder[2].label}
                   </Typography>
                 </Tooltip>
                 <Tooltip title={sectionsOrder[2].title}>
@@ -493,17 +455,16 @@ const Home = () => {
                     sx={{
                       cursor: "pointer",
                       borderBottom: theme =>
-                        sectionSelected === 2 ? `solid 2px ${theme.palette.common.orange}` : undefined,
+                        sectionSelected === 3 ? `solid 2px ${theme.palette.common.orange}` : undefined,
                     }}
                     onClick={() => switchSection(2)}
                   >
-                    {sectionsOrder[2].label}
+                    {sectionsOrder[3].label}
                   </Typography>
                 </Tooltip>
               </>
             )}
           </Stack>
-
           {!isMovil && (
             <AppHeaderSearchBar
               searcherUrl={"search"}
@@ -555,27 +516,9 @@ const Home = () => {
           </Stack>
         </Stack>
       </Box>
-      <Box sx={{ height, width: "100%", position: "absolute", top: "0px" }}>
-        <Box
-          sx={{
-            position: "absolute",
-            width: "100%",
-
-            bottom: "0px",
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "center",
-            color: theme => (theme.palette.mode === "dark" ? "#fff" : "#000"),
-          }}
-          className={showLandingOptions ? "show-blurred-text" : "hide-content"}
-        >
-          {height > 500 && "Scroll"}
-          <KeyboardDoubleArrowDownIcon fontSize={width < 900 ? "small" : "medium"} />
-        </Box>
-      </Box>
       <Box sx={{ position: "relative" }}>
         <Box
-          sx={{ position: "absolute", top: "70px", bottom: "0px", left: "0px", minWidth: "10px", maxWidth: "180px" }}
+          sx={{ position: "absolute", top: height, bottom: "0px", left: "0px", minWidth: "10px", maxWidth: "180px" }}
         >
           <Box sx={{ position: "sticky", top: "100px", zIndex: 11 }}>
             <MemoizedTableOfContent
@@ -600,52 +543,35 @@ const Home = () => {
             width: "100%",
             position: "absolute",
             top: 0,
-            background: "#123",
           }}
         >
-          {/* <Typography
-            color="white"
-            variant="h5"
-            sx={{ textAlign: "center" }}
-            className={showLandingOptions ? "show-blurred-text" : "hide-content"}
-          >
-            WHERE WE TAKE NOTES <b>TOGETHER</b>.
-          </Typography>
-          <Button
-            // color="secondary"
-            variant="contained"
-            size={width < 900 ? "small" : "large"}
-            component="a"
-            // href="#JoinUsSection"
-            target="_blank"
-            href="https://1cademy.us/#JoinUsSection"
-            sx={{ minWidth: 200, zIndex: 13, textTransform: "uppercase" }}
-            className={showLandingOptions ? "show-blurred-text" : "hide-content"}
-          >
-            Apply to Join Us!
-          </Button>
           <Box
-            sx={{ display: "flex", flexDirection: "column", alignItems: "center", color: "common.white" }}
+            sx={{
+              display: "flex",
+              flexDirection: "column",
+              alignItems: "center",
+              fontWeight: "700",
+              color: theme => (theme.palette.mode === "dark" ? "#fff" : "#000"),
+            }}
             className={showLandingOptions ? "show-blurred-text" : "hide-content"}
           >
             {height > 500 && "Scroll"}
             <KeyboardDoubleArrowDownIcon fontSize={width < 900 ? "small" : "medium"} />
-          </Box> */}
+          </Box>
         </Stack>
 
         <Box sx={{ width: "100%", maxWidth: "980px", px: isDesktop ? "0px" : "10px", margin: "auto" }}>
-          <Box id={sectionsOrder[1].id} ref={howSectionRef} sx={{ pb: 10, scrollMarginTop: "70px" }}>
+          <Box id={sectionsOrder[1].id} ref={howSectionRef} sx={{ pb: 10 }}>
             <HowItWorks
               section={sectionSelected}
               // ref={sectionAnimationControllerRef}
-              //   artboards={[...section1ArtBoards, ...artboards]}
-              artboards={artboards}
+              artboards={[...section1ArtBoards, ...artboards]}
             >
               <Box sx={{ position: "relative", width: "inherit", height: "inherit" }}>
-                <RiveComponent0 className={`rive-canvas ${idxRiveComponent !== 0 ? "rive-canvas-hidden" : ""}`} />
-                <RiveComponent1 className={`rive-canvas ${idxRiveComponent !== 1 ? "rive-canvas-hidden" : ""}`} />
-                <RiveComponent2 className={`rive-canvas ${idxRiveComponent !== 2 ? "rive-canvas-hidden" : ""}`} />
-                <RiveComponent3 className={`rive-canvas ${idxRiveComponent !== 3 ? "rive-canvas-hidden" : ""}`} />
+                <RiveComponent1 className={`rive-canvas ${idxRiveComponent !== 0 ? "rive-canvas-hidden" : ""}`} />
+                <RiveComponent2 className={`rive-canvas ${idxRiveComponent !== 1 ? "rive-canvas-hidden" : ""}`} />
+                <RiveComponent3 className={`rive-canvas ${idxRiveComponent !== 2 ? "rive-canvas-hidden" : ""}`} />
+                <RiveComponent4 className={`rive-canvas ${idxRiveComponent !== 3 ? "rive-canvas-hidden" : ""}`} />
               </Box>
             </HowItWorks>
           </Box>
@@ -658,7 +584,7 @@ const Home = () => {
               align="center"
               sx={{ pb: 10, fontWeight: 700 }}
             >
-              {sectionsOrder[1].title}
+              {sectionsOrder[2].title}
             </CustomTypography>
             {!whyInViewOnce && <div style={{ height: 2 * height /* background: "red" */ }}></div>}
             {whyInViewOnce && (
@@ -683,61 +609,6 @@ const Home = () => {
             )}
           </Box>
 
-          {/* <Box id={sectionsOrder[3].id} ref={whatSectionRef} sx={{ py: 10 }}>
-            <CustomTypography
-              component={"h2"}
-              variant="h1"
-              marked="center"
-              align="center"
-              sx={{ pb: 10, fontWeight: 700 }}
-            >
-              {sectionsOrder[3].title}
-            </CustomTypography>
-            {!whatInViewOnce ? (
-              <div style={{ height: 2 * height  }}></div>
-            ) : (
-              <Suspense
-                fallback={
-                  <Grid container spacing={1}>
-                    {new Array(8).fill(0).map((a, i) => (
-                      <Grid key={i} item xs={12} sm={6} md={4} lg={3}>
-                        <Skeleton
-                          variant="rectangular"
-                          height={210}
-                          animation="wave"
-                          sx={{ background: "#72727263", maxWidth: 340 }}
-                        />
-                      </Grid>
-                    ))}
-                  </Grid>
-                }
-              >
-                <What />
-              </Suspense>
-            )}
-          </Box> */}
-
-          {/* <Box id={sectionsOrder[4].id} ref={whereSectionRef} sx={{ py: 10 }}>
-            <CustomTypography
-              component={"h2"}
-              variant="h1"
-              marked="center"
-              align="center"
-              sx={{ pb: 10, fontWeight: 700 }}
-            >
-              {sectionsOrder[4].title}
-            </CustomTypography>
-            {!whereInViewOnce ? (
-              <div style={{ height: 2 * height  }}></div>
-            ) : (
-              <Suspense
-                fallback={<Skeleton variant="rectangular" height={490} animation="wave" sx={{ background: gray02 }} />}
-              >
-                <UniversitiesMap theme={"Dark"} />
-              </Suspense>
-            )}
-          </Box> */}
-
           <Box id={sectionsOrder[3].id} ref={whoSectionRef} sx={{ py: 10 }}>
             <CustomTypography
               component={"h2"}
@@ -746,7 +617,7 @@ const Home = () => {
               align="center"
               sx={{ pb: 10, fontWeight: 700 }}
             >
-              {sectionsOrder[2].title}
+              {sectionsOrder[3].title}
             </CustomTypography>
             {!whoInViewOnce ? (
               <div style={{ height: 2 * height /* background: "pink" */ }}></div>
@@ -783,7 +654,6 @@ const Home = () => {
           </Box>
         </Box>
       </Box>
-
       <Modal
         open={open}
         onClose={() => setOpen(false)}
@@ -830,8 +700,8 @@ const Home = () => {
             theme.palette.mode === "dark" ? "rgba(0,0,0,.72)" : theme.palette.common.darkBackground1,
         }}
       />
-      {/* Mobile Searcher */}
-      {openSearch && <SearcherPupUp onClose={() => setOpenSearch(false)} />}
+      {openSearch && isMovil && <SearcherPupUp onClose={() => setOpenSearch(false)} />}
+
       <style>{`
           body{
             overflow:hidden;
@@ -850,6 +720,7 @@ export default Home;
 
 const advanceAnimationTo = (rive: Rive, timeInSeconds: number, theme?: any) => {
   rive.scrub(theme.palette.mode === "dark" ? "dark" : "light", 1);
+
   //@ts-ignore
   if (!rive?.animator?.animations[0]) return;
   //@ts-ignore
